fast-api/requirements.txt
albucore==0.0.24
albumentations==2.0.8
annotated-types==0.7.0
anyio==4.10.0
argon2-cffi==25.1.0
argon2-cffi-bindings==25.1.0
arrow==1.3.0
asgiref==3.10.0
asttokens==3.0.0
async-lru==2.0.5
async-timeout==5.0.1
attrs==25.4.0
babel==2.17.0
beautifulsoup4==4.14.2
bleach==6.2.0
Booktype==1.5
Brotli==1.1.0
certifi==2025.10.5
cffi==2.0.0
charset-normalizer==3.4.4
click==8.2.1
colorama==0.4.6
coloredlogs==15.0.1
comm==0.2.3
contourpy==1.3.3
cssselect2==0.8.0
cycler==0.12.1
Cython==3.1.4
debugpy==1.8.17
decorator==5.2.1
defusedxml==0.7.1
Django==5.2.7
dlib==20.0.0
easydict==1.13
executing==2.2.1
face-recognition==1.3.0
face-recognition-models==0.3.0
facenet-pytorch==2.6.0
fastapi==0.116.1
fastjsonschema==2.21.2
filelock==3.20.0
flatbuffers==25.9.23
fonttools==4.60.1
fqdn==1.5.1
fsspec==2025.9.0
greenlet==3.2.4
h11==0.16.0
httpcore==1.0.9
httpx==0.28.1
humanfriendly==10.0
idna==3.10
imageio==2.37.0
insightface==0.7.3
ipykernel==7.0.1
ipython==9.6.0
ipython_pygments_lexers==1.1.1
isoduration==20.11.0
jedi==0.19.2
Jinja2==3.1.6
joblib==1.5.2
json5==0.12.1
jsonpointer==3.0.0
jsonschema==4.25.1
jsonschema-specifications==2025.9.1
jupyter-events==0.12.0
jupyter-lsp==2.3.0
jupyter_client==8.6.3
jupyter_core==5.8.1
jupyter_server==2.17.0
jupyter_server_terminals==0.5.3
jupyterlab==4.4.9
jupyterlab_pygments==0.3.0
jupyterlab_server==2.27.3
kiwisolver==1.4.9
lark==1.3.0
lazy_loader==0.4
Markdown==3.9
MarkupSafe==3.0.3
matplotlib==3.10.7
matplotlib-inline==0.1.7
mistune==3.1.4
ml_dtypes==0.5.3
mpmath==1.3.0
nbclient==0.10.2
nbconvert==7.16.6
nbformat==5.10.4
nest-asyncio==1.6.0
networkx==3.5
notebook==7.4.7
notebook_shim==0.2.4
numpy==1.26.4
onnx==1.19.1
onnxruntime==1.23.1
opencv-python==4.12.0.88
opencv-python-headless==4.12.0.88
overrides==7.7.0
packaging==25.0
pandas==2.3.3
pandas-stubs==2.3.2.250926
pandocfilters==1.5.1
parso==0.8.5
pillow==10.2.0
platformdirs==4.5.0
prettytable==3.16.0
prometheus_client==0.23.1
prompt_toolkit==3.0.52
protobuf==6.33.0
psutil==7.1.0
pure_eval==0.2.3
pycparser==2.23
pycryptodomex==3.23.0
pydantic==2.11.7
pydantic_core==2.33.2
pydyf==0.11.0
Pygments==2.19.2
pypandoc==1.15
pyparsing==3.2.5
pyphen==0.17.2
pyreadline3==3.5.4
python-dateutil==2.9.0.post0
python-json-logger==4.0.0
python-multipart==0.0.20
pytz==2025.2
pywin32==311
pywinpty==3.0.2
PyYAML==6.0.3
pyzmq==27.1.0
redis==6.4.0
referencing==0.37.0
requests==2.32.5
rfc3339-validator==0.1.4
rfc3986-validator==0.1.1
rfc3987-syntax==1.1.0
rpds-py==0.27.1
scikit-image==0.25.2
scikit-learn==1.7.2
scipy==1.16.2
Send2Trash==1.8.3
simplejson==3.20.2
simsimd==6.5.3
six==1.17.0
sniffio==1.3.1
soupsieve==2.8
SQLAlchemy==2.0.43
sqlite-vec==0.1.6
sqlparse==0.5.3
stack-data==0.6.3
starlette==0.47.3
stringzilla==4.2.1
sympy==1.14.0
terminado==0.18.1
threadpoolctl==3.6.0
tifffile==2025.10.4
tinycss2==1.4.0
tinyhtml5==2.0.0
torch==2.2.2
torchvision==0.17.2
tornado==6.5.2
tqdm==4.67.1
traitlets==5.14.3
types-python-dateutil==2.9.0.20251008
typing-inspection==0.4.1
typing_extensions==4.15.0
tzdata==2025.2
unicorn==2.1.3
uri-template==1.3.0
urllib3==2.5.0
uvicorn==0.35.0
wcwidth==0.2.14
weasyprint==66.0
webcolors==24.11.1
webencodings==0.5.1
websocket-client==1.9.0
zopfli==0.2.3.post1
